Isolate auto extractor types
