+++ src/models/auto-extractor/index.ts
@@ -1,0 +1,12 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export * from './auto-extractor';
+export * from './is-auto-extractor';
+export * from './raw-auto-extractor';
+export * from './to-auto-extractor';
